feat(inventory): split inventory management pages
