+++ _bmad-output/implementation-artifacts/sprint-status.yaml
@@ -1,5 +1,5 @@
 # generated: 2026-04-15
-# last_updated: 2026-05-27 (story-5-7-created)
+# last_updated: 2026-05-27 (story-5-7-review)
 # project: infitek_erp
 # project_key: NOKEY
 # tracking_system: file-system
@@ -39,7 +39,7 @@
 #   _bmad-output/planning-artifacts/flow-quantity-data-lineage.md
 
 generated: 2026-04-15
-last_updated: 2026-05-27 (story-5-7-created)
+last_updated: 2026-05-27 (story-5-7-review)
 project: infitek_erp
 
 project_key: NOKEY
@@ -114,7 +114,7 @@ development_status:
   5-4-发货需求枢纽详情页: done
   5-5-发货需求下游入口与完整作废逻辑: done
   5-6-发货需求待生成采购单状态纠偏: done
-  5-7-库存管理菜单拆分与页面边界整理: ready-for-dev
+  5-7-库存管理菜单拆分与页面边界整理: review
   epic-5-retrospective: done
 
   # ─────────────────────────────────────────────────
