+++ apps/web/src/pages/inventory/components/InventoryAvailabilityTag.tsx
@@ -1,0 +1,9 @@
+import { inventoryTag } from './inventory-page-utils';
+
+export function InventoryAvailabilityTag({ value }: { value: number }) {
+  return (
+    <span className={`inventory-availability-tag ${inventoryTag(value)}`}>
+      {value > 0 ? `可用 ${value}` : '无可用库存'}
+    </span>
+  );
+}
